Added loading usable items list from JSON file.
Implemented but not used.

+++ Assets/Scripts/Items/Features/ContinuesStatModFeature.cs
@@ -3,7 +3,7 @@
     public float Duration { get; private set; }
     public float Interval { get; private set; }
 
-    public ContinuesStatModFeature(float duration, float interval, int amount, StatTypes stat, FeatureTypes type)
+    public ContinuesStatModFeature(float duration, float interval, float amount, StatTypes stat, FeatureTypes type)
         : base(amount, stat, type)
     {
         Duration = duration;
