+++ Assets/Scripts/Items/Features/InstantStatModFeature.cs
@@ -1,5 +1,5 @@
 ﻿public class InstantStatModFeature : Feature {
-    public InstantStatModFeature(int amount, StatTypes stat, FeatureTypes type)
+    public InstantStatModFeature(float amount, StatTypes stat, FeatureTypes type)
         : base(amount, stat, type)
     {
